Removing archaius from hermes (#1508)
* Removed SslContextProperties from Configs

* move consumer receiver configuration to spring properties (#1509)

* move consumer receiver configuration to spring properties
* new approach for configuration


* Added overriding variable in integration tests

* migrate configuration for kafka.consumers (#1512)

* migrate configuration for kafka.consumers


* Migrate configuration for http clients, commit offset and consumer rate, removed unused properties (#1513)

* Migrate configuration for http and http2 client

* Migrate configuration for commit offset properties

* Migrate configuration for sender async timeout

* Migrate configuration for consumer rate

* Code review fixes

* Migrate consumer batch config and google pubsub sender properties (#1514)

* Migrate consumer batch properties

* Migrate google pubsub sender properties

* Migrate consumer configuration for kafka-authorization (#1518)

* Migrate consumer configuration for kafka-consumers


* Migrate consumer wokload and maxrate properties (#1519)

* Migrate consumer wokload properties

* Migrate consumer wokload properties

* Migrate consumer wokload properties

* Fixed integrationTest

* Migrated maxrate config, fixed integrationTest

* Fixed ci

* Fixed ci

* Code review

* Code review

* Migrate common consumer props (#1525)

* Migrate consumer wokload properties

* Migrate consumer wokload properties

* Fixed integrationTest

* Migrated maxrate config, fixed integrationTest

* Fixed ci

* Fixed ci

* Code review

* Code review

* Migrated common consumer configs

* Code review fixes

* Code rewiev fixes

* Changed time config props to duration

* Changed time config props to duration

* Deleted unused code

* Migrated time config props to Duration

* Refactored parameters

* Removed unused branch from github ci

* Code review fixes

* Removing archaius from frontends  (#1526)

* initial frontend spring configuration

* Migrated messagePreview, throughputLimiter and topicLoading properties (#1528)

* Migrated messagePreview and throughputLimiter properties

* Added ci to branch

* Fixed test

* Fixed test

* Migrated TopicLoadingProperties

* Migrated LocalMessageStorage properties (#1529)

* Migrated LocalMessageStorage properties

* Added ci

* Fixed git merge

* Migrated HandlersChain and ReadinessCheck properties (#1530)

* Migrated HandlersChain and ReadinessCheck properties

* Code reviev fix

* Code reviev fix

* Frontend server and ssl properties migrated (#1536)

* Migrated frontend ssl properties

* Migrated hermes server properties

* Fixed benchmarks

* Migrated time config props to Duration

* Fixed benchmarks

* Removing archaius from commons (#1531)

* Initial commit

* Migrated schema properties (#1537)

* Fixed merging bug

* Migrated metrics and graphite properties (#1532)

* Migrated metrics and graphite properties

* Added CI

* Merge commit

* Migrated zookeeper configuration properties (#1533)

* Migrated zookeeper configuration properties

* Added CI

* Migrated zookeeper and kafka properties, fully removed archaius (#1539)

* Migrated kafka header name properties

* Migrated kafka producers properties

* Initial kafka properties migration commit

* kafka properties migration

* kafka properties migration

* Code refactor

* Fully removed archaius

* Refactored consumer kafka properties

* Refactored consumer kafka and zookeeper properties

* Refactored frontend zookeeper properties

* Refactored frontend kafka properties

* Changed yml to yaml format, added logging in DatacenterNameProvider

* Fixed graphite metrics test

* Refactored docker for consumers and frontend

* Refactore Parameters config classes

* Migrated time config props to duration

* Fixed merge


* Reverted additional branch from ci.yml and removed unused property

+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/registry/ConsumerNodesRegistry.java
@@ -41,7 +41,7 @@ public class ConsumerNodesRegistry extends PathChildrenCache implements PathChil
             ExecutorService executorService,
             ConsumerNodesRegistryPaths registryPaths,
             String consumerNodeId,
-            int deathOfConsumerAfterSeconds,
+            long deathOfConsumerAfterSeconds,
             Clock clock
     ) {
         super(curatorClient, registryPaths.nodesPath(), true, false, executorService);
